Extracting pipeline.go
