+++ lib/deployable_chans.go
@@ -1,7 +1,6 @@
 package sous
 
 import (
-	"context"
 	"sync"
 )
 
@@ -74,80 +73,4 @@ func (d *DeployableChans) collect() diffSet {
 // ID returns the ID of this DeployablePair.
 func (dp *DeployablePair) ID() DeploymentID {
 	return dp.name
-}
-
-// DeployableProcessor processes DeployablePairs off of a DeployableChans channel
-type DeployableProcessor interface {
-	Start(dp *DeployablePair) (*DeployablePair, *DiffResolution)
-	Stop(dp *DeployablePair) (*DeployablePair, *DiffResolution)
-	Stable(dp *DeployablePair) (*DeployablePair, *DiffResolution)
-	Update(dp *DeployablePair) (*DeployablePair, *DiffResolution)
-}
-
-// Pipeline attaches a DeployableProcessor to the DeployableChans, and returns a new DeployableChans.
-func (d *DeployableChans) Pipeline(ctx context.Context, proc DeployableProcessor) *DeployableChans {
-	out := NewDeployableChans(1)
-
-	wg := sync.WaitGroup{}
-	wg.Add(4)
-
-	go func() {
-		for rez := range d.Errs {
-			out.Errs <- rez
-		}
-		wg.Wait()
-		close(out.Errs)
-	}()
-
-	process := func(from, to chan *DeployablePair, doProc func(dp *DeployablePair) (*DeployablePair, *DiffResolution)) {
-		defer close(to)
-		defer wg.Done()
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case dp, ok := <-from:
-				if ok {
-					proked, rez := doProc(dp)
-					if rez != nil {
-						out.Errs <- rez
-					}
-					if proked != nil {
-						to <- proked
-					}
-				} else {
-					return
-				}
-			}
-		}
-	}
-
-	go process(d.Start, out.Start, proc.Start)
-	go process(d.Stop, out.Stop, proc.Stop)
-	go process(d.Stable, out.Stable, proc.Stable)
-	go process(d.Update, out.Update, proc.Update)
-	return out
-}
-
-// DeployablePassThrough implements DeployableProcessor trivially: each method simply passes its argument on
-type DeployablePassThrough struct{}
-
-// Start returns its argument
-func (DeployablePassThrough) Start(dp *DeployablePair) (*DeployablePair, *DiffResolution) {
-	return dp, nil
-}
-
-// Stop returns its argument
-func (DeployablePassThrough) Stop(dp *DeployablePair) (*DeployablePair, *DiffResolution) {
-	return dp, nil
-}
-
-// Stable returns its argument
-func (DeployablePassThrough) Stable(dp *DeployablePair) (*DeployablePair, *DiffResolution) {
-	return dp, nil
-}
-
-// Update returns its argument
-func (DeployablePassThrough) Update(dp *DeployablePair) (*DeployablePair, *DiffResolution) {
-	return dp, nil
 }
